Bro.ai/Bro.ai/requirements.txt
pywebview
SpeechRecognition
pyaudio
pyttsx3
pillow
pyautogui
psutil
requests
python-dotenv
langdetect